backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
django-storages
boto3
